requirements.txt
absl-py
aniso8601
annotated-types
anyio
astunparse
blinker
cachetools
charset-normalizer
click
colorama
distro
et-xmlfile
exceptiongroup
Flask
Flask-JWT-Extended
Flask-RESTful
gast
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
groq
grpcio
grpcio-status
h11
h5py
httpcore
httpx
idna
itsdangerous
Jinja2
libclang
Markdown
numpy
oauthlib
openpyxl
opt-einsum
packaging
proto-plus
pyasn1
pyasn1_modules
pydantic
pydantic_core
PyJWT
python-dotenv
pytz
requests-oauthlib
rsa
sniffio
tensorboard
tensorboard-data-server
tensorflow-estimator
tensorflow-hub
tensorflow-io-gcs-filesystem
termcolor
tf-keras
typing_extensions
urllib3
Werkzeug
wrapt
WTForms
zipp
pillow
